pyproject.toml
[tool.poetry]
name = "solar-energy-system-backend"
version = "0.1.0"
description = ""
authors = ["levakuz <levakuz007@gmail.com>"]
readme = "README.md"
packages = [{include = "solar_energy_system_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
mypy = "^1.3.0"
pylint = "^2.17.4"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.3"}
aerich = "^0.7.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
types-passlib = "^1.7.7.12"
types-python-jose = "^3.3.4.7"
python-multipart = "^0.0.6"



[tool.aerich]
tortoise_orm = "src.settings.TORTOISE_ORM"
location = "src/database/migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
